VEA-16 Moved auth DTOs into their own package.
Made a package for auth DTOs (LoginRequest, RegisterRequest,
AuthResponse) and moved them into the package for better organization.
