+++ src/main/java/com/danieldigiovanni/email/auth/dto/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.danieldigiovanni.email.auth;
+package com.danieldigiovanni.email.auth.dto;
 
 import jakarta.validation.constraints.NotNull;
